platform-fru-detect: Detect the platform we're running on
And if it isn't supported, exit cleanly after reporting so.

Detect the platform using the model string from the devicetree. Do this
to minimise cognitive load with respect to the relationships between
platform-fru-detect and other daemons in the system.

As it stands we have the following, where '->' is defined as "provides
data to":

platform-fru-detect -> phosphor-inventory-manager
phosphor-inventory-manager -> entity-manager
entity-manager -> nvmesensor
nvmesensor -> phosphor-fan-control

On P10 platforms, entity-manager provides the IBMCompatibleSystem schema
which other applications can use to determine what configuration files
to load. This was done to mitigate a (potentially large) amount of
application-specific configuration data appearing on D-Bus and requiring
application-specific schemas. It's feasible the IBMCompatibleSystem
schema could be used by platform-fru-detect to determine the platform on
which it is running, however then we have a cyclic set of relationships:

entity-manager -> platform-fru-detect
platform-fru-detect -> phosphor-inventory-manager
phosphor-inventory-manager -> entity-manager

As platform-fru-detect is already using platform-specific tricks to
learn about the presence of undiscoverable FRUs, the harm incurred by
inspecting the devicetree feels minimal.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>