rainier: Untangle Nisqually, Flett and Williwakas
Sorry, this patch is an ugly one, but making it not ugly is more effort
than it's worth. This gets us to a much better place where Williwakas
no-longer depends on a Flett instance. Instead, we split the work of
populating the inventory between two different views of the NVMe drive:

1. The view from Williwakas, which amounts to just drive presence, and
2. The view from Flett, which can communicate with drive endpoints but
   has no means to learn of changes to drive presence

To that end, the NVMeDrive class is split in two, with the original
class remaining as a base class. The two derivatives are
WilliwakasNVMeDrive and FlettNVMeDrive, which have responsibilities as
described above.

The implementation of WilliwakasNVMeDrive and FlettNVMeDrive is pushed
into the Rainier platform implementation as they both deal with
platform-specifics. From there, the types and relationships between the
platform device classes are untangled and the device detection logic
reworked to separate the representation of the two types of cards.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>