platform-fru-detect: Add Notifier and NotifySink
Notifier is a wrapper around epoll(2), and NotifySink is a base class
for classes that wish to be notified of an event which (typically) one
of their own resources has produced.

Basically, improve readability and type-safety by centralising all the
epoll gunk behind something that doesn't amount to passing around raw
file descriptors.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>