platform-fru-detect: Deploy PublishWhenPresentInventoryDecorator
Make sure the inventory state provided by WilliwakasNVMeDrive and
FlettNVMeDrive are seen together. This ensures nvmesensor is only
triggered when a drive is actually present.

This is a work-around that's required until entity-manager can be taught
to respect the Present property on xyz.openbmc_project.Inventory.Item,
though this is kind of counter to the entire design-point of
entity-manager.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>