Enable/dis PCIe card TMP435s on slot power changes
The TMP435 devices on the Flett and Bear Lake/River PCIe cable cards are
connected to PCIe slot power, not main power, which doesn't get asserted
until the hypervisor turns on the slot power after they start running.

This means that the device driver can't probe the device until then. To
support this, start watching the PowerState D-Bus property on the
xyz.openbmc_project.State.Decorator.PowerState interface on the PCIe
slot D-Bus objects, which will be set to 'On' by PLDM when it is told to
do so by the hypervisor.

The Bear Lake/River cable cards have the same TMP435 address mappings as
the Fletts, but can be in more slots - all but C6.  To detect these
cards, their known model values will be compared against the model value
in the inventory.

The existing requireDevice() function will be used to ensure the driver
is bound when the card is powered on, with releaseDevice() being used
when the card is powered off.

The InventoryManager class has the sdbusplus match object for the
propertiesChanged signal, and it tells the platform manager which tells
the loaded platform about the change.  The platform then tells Nisqually
about it.

There will need to be a bit of refactoring in the future when Everest is
added so more common code can be used, but I'll wait until the NVME
support for Everest is added so I can see what the framework will look
like.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>