Check for powered on PCIe slots on startup
If PCIe slots are already powered on when the application starts, it's
most likely because the BMC rebooted at runtime so the drivers need to
be enabled again.  Find and check the power state on all the PCIe slots
for this case, and do the driver enables if necessary.

The code can handle the case where the driver was already enabled,
which could happen if only the application restarts at runtime.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>