inventory: Begin inventory refactor with Interface class
Interface represents some arbitrary D-Bus interface that is implemented
on inventory objects in phosphor-inventory-manager. It's implementation
is such that it captures the details necessary for adding and "removing"
the interface data (removing an interface in PIM is implemented by
eliding the data in the interface's properties) without requiring access
to a specific subclass.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>