nvme: Push inventory knowledge out of core drive types
Enable the switch from inheritance to composition for NVMeDrive objects.
Specifically, with this change, BasicNVMeDrive still requires the
physical presence of the hardware to function, but can be instantiated
independently of the inventory information associated with the drive.
This leaves the containing object free to be instantiated to cater to
the inventory requirements of the drive without probing the drive
itself. This in-turn is important for cleaning up the inventory after
lights-out unplug events.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>