inventory: Introduce migrations
Migrations allow us to clean up the inventory in the case of misplaced
interfaces. We use migrations to separate the fixup code from the code
managing the interface values under normal operation in an effort to
keep things tidy. As such, migrations are implemented as visitor objects
applied across all inventory objects as fetched by the inventory
implementation.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>