inventory: Fix PublishWhenPresent cache management
Inventory operations on the PublishWhenPresentInventoryDecorator were
sensitive to ordering. Specifically, attempting to mark an NVMe drive as
removed when the first event is the discovery that it's not reachable
over the I2C interface rather than the GPIO presence line lead to the
failure to appropriately update the inventory.

Rework the object cache to take values in terms of a map keyed by the
interface name whilst still holding values of type
inventory::interfaces::Interface. This makes it trivial to replace the
remove-only construction in the add() path with an instance that
contains the defined 'add' property values.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>