platforms: Handle objects that may not be defined in inventory
In some cases the hardware schematics allow for more devices to be
populated than the system configuration policy allows. The application
is implemented in terms of the schematics and so allows for population
of this hardware, but other firmware components such as the inventory
database may instead follow the system design policy. In such cases,
catch the exceptions thrown when attempting to update the inventory.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>